update.sh: prevent apache out of memory
With a default number of workers set to 150, the example apache
configuration will use over 4GB of RAM in the following, rather
likely, demonstration scenario:

* sync a directory with 1000 photos in one directory from a desktop to
  nextCloud
* go to the web interface and quickly scroll down to the bottom of the
  list of photos that were just uploaded
* apache will run as many apache workers as possible (150) to create
  thumbnails and use over 4GB of RAM

The default number of workers is reduced to 30 which has been
experienced to fit in a 2GB virtual machine. It takes longer to
complete but does not trigger out of memory problems.