Removes VOLUME directive.
In a kubernetes env, this is causing a lot of issues.
We want a read only docker image. And this helps.
And as you see it removes a lot of logic that is always a good things to remove lines ;)

The last logic piece that is needed, is to know if the Nc is installed or not, and we can get this from `occ status`.

In term of upgrade Nobody will lost the data.
There is just one thing, for people relying on /var/www/html/ is that the Nextcloud will not update anymore.
But I guess people will realize and search what is happening.
For convenience, we could put a warning in the entrypoint, and detect this case.
What do you think? I know it is tricky to update these things, but I think we agree that we have to get rid of this old logic.

Signed-off-by: pierreozoux <pierre@ozoux.net>