Update docker-entrypoint.sh
Added a helper function to check whether the upgrade is doable: returns true if major version (installed) + 1 is -lt major version (image).
This lets you upgrade from eg. 10.X.Y -> 10.X+1.Z and 10.X.Y -> 11.A.B, but not 10.X.Y -> 12.A.B.
If it detects this, the script will notify and bail immediately.