README.md: clarify that _HOST variables cannot be files
The auto configuration sections seems to imply that POSTGRES_HOST and
MYSQL_HOST can be file variables, while this is not supported. If all
variables are set as files including the host, the installation will
not be done automatically.
This can be hard to figure out, as during the first start of the
container nothing indicates why the auto configuration is not done.

To try to make it a bit clearer in the README, explicitly mention that
using a file for the hosts variable is not supported.

Furthermore, in the docker secrets section it's currently easy to
misread the fact that the last variable mentioned (the host) is not
using '_FILE'. Since it's anyway not part of the group of variables
that support secrets, remove it from the list to avoid confusion.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>