Improve handling of secret variables and files in autoconfig
In the recent state, the autoconfig feature required either all database
secrets to be passed as environment variables, or all to be passed as
secret files, but didn't allow to use a mix of these variants, e.g.
passing the database user and name as environment variables, and only
the password as a file.

This is now fixed and allows working with database values the same way,
as the Postgres and MySQL/MariaDB images do; with mixed variants also.

Signed-off-by: Patrick Hobusch <patrick@hobusch.net>