Update Dockerfile
Suggesting (#line 101) to improve opcache.memory_consumption to value 256 instead of 128, in order to avoid the following recurrent warning, even when a rather light Nextcloud installation, with very few (4/5) apps is used.
```
"The PHP OPcache module is not properly configured. The OPcache buffer is nearly full. To assure that all scripts can be hold in cache, it is recommended to apply "opcache.memory_consumption" to your PHP configuration with a value higher than "128".
```

Signed-off-by: Spyderfoo <36179436+neokyx@users.noreply.github.com>